FIX: return src_str from is_invalid()
Before the '(#7) FIX 1.' is_invalid() would first edit out all the whitespaces in src_str and riassign it to the lexer
object after all the checks. '(#7) FIX 1.' introduced a bug: is_invalid() was getting called first but after all the
checks the edited src_str was NOT getting assigned to the object; instead the un-edited src_str arg from the constructor
was the one getting assigned to the object, which would of course mess up the tokenization. Now all of that has been
compressed by making is_invalid() return and assign the edited src_str to the lexer object.